page time index
close #5

+++ utils/utils.cpp
@@ -28,10 +28,12 @@ bool utils::rm(const std::string&rm_path){
     if(!boost::filesystem::exists(rm_path))
         return true;
 	try {
-		boost::filesystem::path path_to_remove(rm_path);
-		for (boost::filesystem::directory_iterator end_dir_it, it(path_to_remove); it != end_dir_it; ++it) {
-			if (!boost::filesystem::remove_all(it->path())) {
-				return false;
+		if (boost::filesystem::is_directory(rm_path)) {
+			boost::filesystem::path path_to_remove(rm_path);
+			for (boost::filesystem::directory_iterator end_dir_it, it(path_to_remove); it != end_dir_it; ++it) {
+				if (!boost::filesystem::remove_all(it->path())) {
+					return false;
+				}
 			}
 		}
 		boost::filesystem::remove_all(rm_path);
